+++ app/controllers/projects_controller.rb
@@ -87,7 +87,7 @@ class ProjectsController < ApplicationController
   def project_params
     params.require(:project).permit(
       :name, :slug, :description, :disable_all_emails,
-      :collect_email_on_surveys, :hide_values_on_pdfs,
+      :hide_values_on_pdfs,
       :randomizations_enabled, :adverse_events_enabled, :blinding_enabled,
       :handoffs_enabled, :auto_lock_sheets,
       # Uploaded Logo