+++ db/migrate/20160929183822_remove_collect_email_on_surveys_from_projects.rb
@@ -1,0 +1,5 @@
+class RemoveCollectEmailOnSurveysFromProjects < ActiveRecord::Migration[5.0]
+  def change
+    remove_column :projects, :collect_email_on_surveys, :boolean, null: false, default: true
+  end
+end