+++ db/migrate/20160929185304_remove_email_from_subjects.rb
@@ -1,0 +1,5 @@
+class RemoveEmailFromSubjects < ActiveRecord::Migration[5.0]
+  def change
+    remove_column :subjects, :email, :string
+  end
+end